refactor: Choose operation through parameter
